Updated the code to 2020
